+++ utils/accessibility/models/A11yError.ts
@@ -1,0 +1,12 @@
+import { Target } from './Target';
+
+export interface A11yError {
+    id: string;
+    description: string;
+    wcagRule: string;
+    severity: string;
+    help: string;
+    helpUrl: string;
+    guideline: string;
+    target: Target[];
+}