+++ utils/accessibility/models/ErrorByLevel.ts
@@ -1,4 +1,0 @@
-export interface ErrorByLevel {
-    page: string;
-    total: number;
-}