+++ utils/accessibility/models/ImpactType.ts
@@ -1,7 +1,0 @@
-export enum ImpactType {
-    minor = '#0071ff',
-    moderate = '#CF30CB',
-    serious = '#63954a',
-    critical = '#ff0019',
-    noImpact = '#000'
-}