+++ utils/accessibility/models/Report.ts
@@ -1,0 +1,8 @@
+import { A11yError } from './A11yError';
+
+export interface ReportData {
+    pageKey: string;
+    accessibilityScore: number;
+    video: string;
+    errors: A11yError[];
+}