+++ utils/accessibility/models/Target.ts
@@ -1,0 +1,4 @@
+export interface Target {
+    element: string,
+    screenshot: string
+}